Improved Compiler semantic analyzer - step 1
The existing semantic analysis was designed purely to support live semantic
warnings in the code browser, e.g. shadowing inst vars with temps, read-
before-written, unreferenced locals, etc. Much of this analysis overlaps
with that required to implement the compiler back end in Smalltalk, but
needs to be richer for that purpose. Some modifications are made to the
parser as it is possible to identify some classes of variable syntactically,
for example we can distinguish between temps and args, and also the pseudo
variables "self", "super", and "thisContext". Identifying whether variable
references are to instance variables, pool variables, or temporary variables,
and determining when local variables are copied or closed over, is the
responsibility of the semantic analyser.

Todo:
1) Lexical analysis to identify copied and context temps and copying/full
blocks. This will be similar to the lexical analysis implemented in the
existing Dolphin (C++) compiler.
2) Parse node visitor to transform the AST to logical instructions (an
intermediate representation, or IR).
3) Optimization of the IR.
4) Generation of bytecodes from the IR.